Extract file content normalization partial
